+++ properties/Adapter/ReAddingFilesHasNoSideEffect.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Properties\Innmind\Filesystem\Adapter;
 
-use Innmind\Filesystem\Adapter;
+use Innmind\Filesystem\{
+    Adapter,
+    Directory,
+};
 use Innmind\BlackBox\{
     Property,
     Set,
@@ -33,6 +36,11 @@ final class ReAddingFilesHasNoSideEffect implements Property
             ->foreach(static function($file) use ($assert, $adapter) {
                 $adapter->add($file);
                 $assert->true($adapter->contains($file->name()));
+
+                if ($file instanceof Directory) {
+                    return;
+                }
+
                 $assert->same(
                     $file->content()->toString(),
                     $adapter