+++ src/Adapter.php
@@ -13,10 +13,10 @@ use Innmind\Immutable\{
  */
 interface Adapter
 {
-    public function add(File $file): void;
+    public function add(File|Directory $file): void;
 
     /**
-     * @return Maybe<File>
+     * @return Maybe<File|Directory>
      */
     public function get(Name $file): Maybe;
     public function contains(Name $file): bool;