+++ src/Adapter/Logger.php
@@ -31,7 +31,7 @@ final class Logger implements Adapter
         return new self($filesystem, $logger);
     }
 
-    public function add(File $file): void
+    public function add(File|Directory $file): void
     {
         $this->logger->debug('Adding file {file}', ['file' => $file->name()->toString()]);
         $this->filesystem->add($file);